bld: Standarize unit tests for all backends

+++ src/nonim/backend/test/cases/expression_indexed/expected.zig
@@ -1,3 +1,3 @@
-fn get_element (arr: [10]i64, idx: isize) isize {
+fn get_element (arr: [10]isize, idx: isize) isize {
   return arr[idx];
 }
